[feat] admin开发 报警统计开发

+++ dlmws-admin-backend/admin-main/src/main/java/com/yipingjian/dlmws/warn/entity/Rule.java
@@ -12,7 +12,7 @@ public class Rule {
     private String logType;
     private String keyword;
     private String type;
-    private Integer threshold;
+    private Double threshold;
     private Integer intervalTime;
     private String dingTalkId;
     private String email;
